Revert "Fields: Add a trash quick action for drafts"
This reverts commit c4b0676f911f40a5ce453c2fbb35c9829ec52e0b.

Giving drafts a confirmation-free "Trash" quick action next to the
confirming "Trash…" action of the other statuses turned out to be
jarring: the two labels are too close for users to tell one runs
immediately and the other asks first, and bulk selections mixing
drafts and published posts ended up with a "Trash…" action that only
applied to some of the selected items.

Go back to a single `trashPost` action that asks for confirmation for
every status, drafts included. It stays a secondary action, as it opens
a dialog. The shared `trash-items.ts` helper is removed along with the
only other action that used it.
