+++ routes/post-list/stage.tsx
@@ -228,7 +228,6 @@ function PostListView( {
 			// Clean up URL when delete actions are performed
 			if (
 				actionId === 'move-to-trash' ||
-				actionId === 'move-draft-to-trash' ||
 				actionId === 'permanently-delete'
 			) {
 				cleanupDeletedPostIdsFromUrl( items );
