+++ routes/template-list/stage.tsx
@@ -176,7 +176,6 @@ function TemplateList() {
 			// Clean up URL when delete actions are performed
 			if (
 				actionId === 'move-to-trash' ||
-				actionId === 'move-draft-to-trash' ||
 				actionId === 'permanently-delete'
 			) {
 				cleanupDeletedPostIdsFromUrl( items );
